Fix ci
Lap package installation is unstable and should be upgraded.
With the current version, lap requires to preinstall numpy. Moreover, for an unknown reason, with the python
set up by the default github action, lap install fails even with numpy installed (failing to find it..)

We fix this by installing everything in a venv.